code/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==3.1.0
Flask-WTF==1.2.2s
psycopg2==2.9.3
python-dotenv==0.21.0
Werkzeug==2.3.0
Jinja2==3.1.2
Flask-Login==0.6.3
Flask-Bcrypt==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Babel==4.0.0
